chore: .test.ts, fix imports
